net: shell: Add tcp command
Move "tcp" command to a separate file.
